Removed a few more warnings. Adapted the bisimulation reduction algorithm, such that it spends less time on making vectors containing transitions, by reserving enough space initially for the vectors.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8529 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/data/source/detail/rewrite/jitty.cpp
@@ -338,7 +338,7 @@ static ATermList create_strategy(ATermList rules, ATermAppl jitty_true)
         {
           if ( bs[i] && !used[i] )
           {
-            assert(i<1<<(8*sizeof(int)-1));
+            assert(i<((size_t)1)<<(8*sizeof(int)-1));  // Check whether i can safely be translated into an int.
             deps = ATinsert(deps,(ATerm) ATmakeInt((int)i));
             args[i] += 1;
 //fprintf(stderr,"dep of arg %i\n",i);
@@ -378,7 +378,7 @@ static ATermList create_strategy(ATermList rules, ATermAppl jitty_true)
 
       for (size_t i = 0; i < arity; i++)
       {
-        assert(i<1<<(8*sizeof(int)-1));
+        assert(i<((size_t)1)<<(8*sizeof(int)-1));
         if ( args[i] > max )
         {
           maxidx = (int)i;
